% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allspecs.R
\name{allspecs}
\alias{allspecs}
\title{<allspecs>}
\usage{
allspecs(garchmodels = c("sGARCH", "eGARCH", "gjrGARCH", "GAS"),
  distributions = c("norm", "std", "ged"))
}
\description{
<compute a Bayesian Slot Machine of MS-GARCH specifications>
}
