% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assetbetas.R
\name{assetbetas}
\alias{assetbetas}
\title{<assetbetas>}
\usage{
assetbetas(specm, wspecm = list(), dates, Y, k = 2, p = 1, npar, df, wdf,
  Aic = -1500)
}
\arguments{
\item{specm}{A formula object for the MS conditional mean model to fit.}

\item{wspecm}{A list of formulas for the weighted GLM models to fit (alternative to MS).}

\item{dates}{A Date vector.}

\item{Y}{A vector containing independent variable.}

\item{k}{Integer, number of regimes to fit, default is 2.}

\item{p}{Integer, order of lags to add in the MS specification, default is 1.}

\item{npar}{Integer, number of coefficients to estimate + 1.}

\item{df}{A data.frame object containing regressors and independent variable.}

\item{wdf}{A "verify" object produced by BMLgarch function output.}

\item{Aic}{A real number indicating the maximum desired AIC value.}
}
\description{
<compute conditional mean equations of MSGARCH models>
}
