% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BacktestPORTFOLIOS.R
\name{BacktestPORTFOLIOS}
\alias{BacktestPORTFOLIOS}
\title{<BacktestPORTFOLIOS>}
\usage{
BacktestPORTFOLIOS(scenario.set, cor.matrices = list(), nroll = 90,
  q = 0.025, shortsell = FALSE, max.weight = 1)
}
\arguments{
\item{scenario.set}{A data.frame object containing all the assets in the market.}

\item{cor.matrices}{A list of conditional correlation matrices.}

\item{nroll}{Integer, width of the rolling window.}

\item{q}{real number, quantile for VaRc estimates.}

\item{shortsell}{Logic, if TRUE negative proportions are allowed in the optimized portfolios.}

\item{max.weight}{real number, total capital to invest}
}
\description{
<Rolling window backtesting of optimized portfolios>
}
