% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMLgarchPIT.R
\name{BMLgarch}
\alias{BMLgarch}
\title{<BMLgarch>}
\usage{
BMLgarch(X, dates, k = 2, show.spec = TRUE, burns = 800, mcmc = 1700,
  Dic = -5000, maxht = 2, fortime = 7, ndraws = 700, sds = 3,
  pred.diag = FALSE)
}
\arguments{
\item{X}{A vector containing dependent variable observations.}

\item{k}{Integer, number of regimes to fit in the MSGARCH model, default is 2.}

\item{show.spec}{Logical, if true it shows specifications fitted at each iteration.}

\item{burns}{Integer. MCMC parameter, draws to burn, default is 800.}

\item{mcmc}{Integer. MCMC parameter, draws to keep, default is 1700.}

\item{Dic}{Real number, the maximum value of DIC to accept a model, default is -5000.}

\item{maxht}{Real number, the maximum value of fitted h(t) to keep the model, default value is 2.}

\item{fortime}{Integer, number of n steps ahead to compute out-of-sample forecasts, default is 7.}

\item{ndraws}{Integer, number of MCMC draws to compute at each of the n steps ahead to forecast, default is 700.}

\item{sds}{Real number, upper and lower bounds (in terms of standard deviations) of predicted values,default is 3.}
}
\description{
<Iteratively fitting MSGARCH models relying on their DIC values>
}
