% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutdata.R
\name{cutdata}
\alias{cutdata}
\title{<cutdata>}
\usage{
cutdata(df, nas = 300)
}
\arguments{
\item{df}{A data.frame object.}

\item{nas}{Integer, maximum number of NA's to tolerate at each column.}
}
\description{
<removes columns in a data.frame if the total number of missing values exceed a certain treshold>
}
