% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DynamicCOR - Copia (3).R
\name{DynamicCOR}
\alias{DynamicCOR}
\title{<DynamicCOR>}
\usage{
DynamicCOR(htfactors, rotated)
}
\arguments{
\item{htfactors}{a matrix containing the fitted conditional variance values of factors, estimated with PCA.}

\item{rotated}{Matrix of the eigenvectors for each selected factor.}
}
\description{
<Starting from OGARCH fitted values, computes a matrix time series object (a list) containing dynamic correlation matrices>
}
