% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorforest.R
\name{factorforest}
\alias{factorforest}
\title{<factorforest>}
\usage{
factorforest(factors, data, ntr = 800, window = 100)
}
\arguments{
\item{factors}{Data.frame object containing time series of the estimated factors.}

\item{data}{Data.frame object containing prices time series of each asset in the market.}

\item{ntr}{Integer, number of observations of the training dataset.}

\item{window}{Integer, number of observations to see in the diagnostic plots.}
}
\description{
<Performing and cross-validating random forest models on each asset included in a data.frame object, given a matrix of risk factors obtained with PCA>
}
