% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImputeBBridge.R
\name{imputeBBridge}
\alias{imputeBBridge}
\title{<imputeBBridge>}
\usage{
imputeBBridge(df, meanwindow = 90, verbose = FALSE)
}
\arguments{
\item{df}{A data.frame object containing missing values.}

\item{meanwindow}{Integer, time interval width to compute local mean if the first value of a series is NA, default is 90.}

\item{verbose}{Logical, if TRUE the initial and final value of fitted Brownian Bridges are printed at each iteration, default is FALSE.}
}
\description{
<imputes a brownian bridge for each missing value in a dataset>
}
