% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCAportfolioss.R
\name{PCAportfolios}
\alias{PCAportfolios}
\title{<PCAportfolios>}
\usage{
PCAportfolios(selection, Center = T, Scale = F, ncomp = 6)
}
\arguments{
\item{selection}{A data.frame object.}

\item{Center}{Logical, whether to center data to the mean or not, default is TRUE.}

\item{Scale}{Logical, whether to standardize data or not, default is FALSE.}

\item{ncomp}{Integer, number of principal components (or factors) to retain, default is 6.}
}
\description{
<Performing PCA on financial data and revealing the composition of each factor-representing portfolio.>
}
