% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCAvariables - Copia.R
\name{PCAvariables}
\alias{PCAvariables}
\title{<PCAvariables>}
\usage{
PCAvariables(df, n = 1, minexpl = 0.15)
}
\arguments{
\item{df}{A data.frame object.}

\item{n}{Integer, number of PCA's to simulate in order to impute missing data, with missMDA package, default is 1.}

\item{minexpl}{Real number from 0 to 1, minimum proportion of explained variance to keep each column of the data.frame; default is 0.15.}
}
\description{
<Performing PCA and reducing a data frame dimension by discarding the less explanatory variables in terms of global variance.>
}
