% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yuimaCOGARCH.R
\name{yuimaCOGARCH}
\alias{yuimaCOGARCH}
\title{<yuimaCOGARCH>}
\usage{
yuimaCOGARCH(X, p = 1, q = 1, term = 1600, num = 24000, gamma = TRUE)
}
\description{
<estimation and simulation of a COGARCH(1,1) model on input price data>
}
